digital_inclusion/requirements.txt
Django==2.0.3
psycopg2==2.7.4
gunicorn==19.7.1
django-widgets-improved==1.5.0
Pillow==5.1.0
django-phonenumber-field==2.0.0
django-import-export==1.0.0
requests==2.18.1
